requirements.txt
colorama
pygame
google-cloud-texttospeech
